move support classes to utils.py and update README with instructions to start the bot
