+++ coco/README.md
@@ -19,11 +19,10 @@ This repository contains the source code for a collaborative CoCobot session. It
         "send_command": true
     }
     ```
-    **To skip all this setup above, you could run a [setup script](https://github.com/clp-research/slurk-bots/blob/ccbts/ccbts/setup.sh).** 
-    1. Make sure that the [slurk](https://github.com/clp-research/slurk) and slurk-bots repositories live next to each other on the same level.
-    2. Navigate to the base directory of this repository and run the script to launch this bot, your command should look like ```bash ccbts/setup.sh``` 
-    This script will build and run the docker images, it will initialise all the env variables with the right permissions and it will set up two bots that can talk to each other locally on your computer. The bot will appear in your containers list as ```slurk/ccbts-bot```. At the end of the run there will be two tokens printed in the shell that you will need to paste to access the waiting rooms. 
-5. Save the generated tokens!
+ 4. Make sure that the [slurk](https://github.com/clp-research/slurk) and slurk-bots repositories live next to each other on the same level.
+ 5. Navigate to the base directory of the slurk-bots repository and run the script to launch this bot, your command should look like this:  
+ ```$ python start_bot.py --users 2 --dev coco/ --extra-args coco/args.ini --copy-plugins --waiting-room-layout-dict concierge/waiting_room_layout.json```.  
+ This script will build and run the docker images, it will initialise all the env variables with the right permissions and it will set everything up for testing locally on your computer. The bot will appear in your containers list as ```slurk/coco```.
 
 Note: Every time a new terminal session is started, the env variables will need to be set up again. You can just run the script again.
